proper clean

+++ sandbox/clean.sh
@@ -4,8 +4,14 @@ pwd=`pwd`
 
 shopt -s extglob
 cd $pwd/go && rm -rf api doc include lib misc test AUTHORS CONTRIBUTORS favicon.ico LICENSE PATENTS README robots.txt
-cd $pwd/go/bin && rm !(go)
+cd $pwd/go/bin && rm -f !(go)
+
+# pkg dir
+cd $pwd/go/pkg && rm -rf obj
 for dir in $pwd/go/pkg/*/
 do
-	cd $dir && rm -rf !(runtime.a)
+	b=`basename $dir`
+	if [ ! "$b" = "tool" ]; then
+		cd $dir && rm -rf !(runtime.a)
+	fi
 done
